Player vs AI tictactoe app
So I followed this tutorial, and finished with no bugs. However, there is something wrong with the xml files I think because the app does not display the Xs or Os. Even though it is blank, if you click randomly it will actually place Xs and Os and tell you who is the winner. Run it in the emulator and you'll see what I mean. If you guys know what the issue is, let me know. I'm still trying to figure it out.

Here is the resource I used:
https://ssaurel.medium.com/learn-to-create-a-tic-tac-toe-game-for-android-82c7bf2369de